# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: INDOPACOM IRONVEIN (Resilient Operational Network for Vital Expeditionary Inventory Nodes) API for medical logistics decision support.
servers:
  - url: /api
...
paths:
...
  /mfa/enroll/verify:
    post:
      operationId: verifyMfaEnrollment
      tags: [mfa]
      summary: Verify the first TOTP code, finalize enrollment, return recovery codes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MfaCodeRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MfaEnrollVerifyResponse" }
        "401":
          description: Bad code or no enrollment in progress.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MfaErrorEnvelope" }
  /mfa/verify:
    post:
      operationId: verifyMfa
      tags: [mfa]
      summary: Verify a TOTP code (or recovery code) for the current session
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MfaCodeRequest" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MfaVerifyResponse" }
        "401":
          description: Bad code.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MfaErrorEnvelope" }
        "429":
          description: Account temporarily locked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MfaErrorEnvelope" }
...
components:
  schemas:
...
    MfaCodeRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 16
          description: "6-digit TOTP code from authenticator, or single-use recovery code."
    MfaEnrollVerifyResponse:
      type: object
      required: [enrolled, recoveryCodes]
      properties:
        enrolled: { type: boolean, const: true }
        recoveryCodes:
          type: array
          items: { type: string }
          minItems: 8
          maxItems: 16
          description: "One-time recovery codes — shown ONCE."
    MfaVerifyResponse:
      type: object
      required: [verified, sessionExpiresAt]
      properties:
        verified: { type: boolean, const: true }
        sessionExpiresAt: { type: string, format: date-time }
...
    MfaErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: { type: string }
        detail: { type: string }
        retryAfterSeconds: { type: integer }
...